requirements.txt
streamlit
pandas
numpy
matplotlib
Pillow
opencv-python
tensorflow
scikit-learn
kaggle
opencv-python-headless
Pillow